Split models file into several files
